[#8] Offload key value store data storage to protobuf
Instead of manually copying and verifying equality between the rust
object and the protobuf object, this commit makes the rust object use
the protobuf object as a single source of truth.

When serializing into a file, we shall simply get the protobuf object
and write it. When deserealizing from a file, we shall simply create a
wrapper around the parsed protobuf object in the file.