[#6] Simple ping support from KVSClient to KVSServer
This commit allows the key value store client to send a simple ping to
the key value store server, which handles it and sends a reply back,
while printing the ping message to stdout.

The key value store client does not support persistent connections yet,
and so the server is essentially screaming into the void.