[#6] Basic create message handling in server
This commit allows the server to handle a basic create message from the
key value store client. In order to allow sharing of resources between
closures and contexts, we now use Arc and RwLock to protect and point to
the server and its store.

The Arc allows the store to not be consumed by the tokio::spawn runtime
and instead passes a clone of the Arc. Similarly, the key value store is
protected by a RwLock to handle concurrent creates and reads, as well as
an Arc for lifetime management.